Move DNR module side-effects into a component.
