+++ shared/js/background/background.js
@@ -27,6 +27,7 @@ import ToggleReports from './components/toggle-reports'
 import TrackersGlobal from './components/trackers'
 import DebuggerConnection from './components/debugger-connection'
 import Devtools from './components/devtools'
+import DNR from './components/dnr'
 import initDebugBuild from './devbuild'
 import initReloader from './devbuild-reloader'
 import tabManager from './tab-manager'
@@ -74,6 +75,7 @@ if (BUILD_TARGET === 'chrome' || BUILD_TARGET === 'chrome-mv2') {
 // MV3-only components
 if (BUILD_TARGET === 'chrome') {
     components.scriptInjection = new MV3ContentScriptInjection()
+    components.dnr = new DNR({ settings, tds })
 }
 console.log(new Date(), 'Loaded components:', components)
 // @ts-ignore
