+++ shared/js/background/events.js
@@ -54,32 +54,6 @@ async function onInstalled (details) {
         experiment.setActiveExperiment()
     }
 
-    if (manifestVersion === 3) {
-        await settings.ready()
-
-        // remove any orphaned session rules (can happen on extension update/restart)
-        await flushSessionRules()
-
-        // check that the dynamic rule state is consistent with the rule ranges we expect
-        clearInvalidDynamicRules()
-
-        // create ATB rule if there is a stored value in settings
-        ATB.setOrUpdateATBdnrRule(settings.getSetting('atb'))
-
-        // Refresh the user allowlisting declarativeNetRequest rule, only
-        // necessary to handle the upgrade between MV2 and MV3 extensions.
-        // TODO: Remove this a while after users have all been migrated to
-        //       the MV3 build.
-        const allowlist = settings.getSetting('allowlisted') || {}
-        const allowlistedDomains = []
-        for (const [domain, enabled] of Object.entries(allowlist)) {
-            if (enabled) {
-                allowlistedDomains.push(domain)
-            }
-        }
-        await refreshUserAllowlistRules(allowlistedDomains)
-    }
-
     // Inject the email content script on all tabs upon installation (not needed on Firefox)
     // FIXME the below code throws an unhandled exception in MV3
     try {
